+++ Assets/Scripts/Skills/Condition/SelfHealthCondition.cs
@@ -1,10 +1,10 @@
 using UnityEngine;
 [CreateAssetMenu(fileName = "SelfHealthCondition", menuName = "Condition/SelfHealthCondition")]
 public class SelfHealthCondition : Condition {
-    public float minValue;
-    public float maxValue;
+    public double minValue;
+    public double maxValue;
     public override bool CheckIfMet(Character actor, Character target) {
-        float healthPercentage = actor.Health / actor.characterData.maxHealth * 100;
+        double healthPercentage = actor.Health / actor.characterData.maxHealth * 100;
         return healthPercentage >= minValue && healthPercentage <= maxValue;
     }
 }
