+++ Assets/Scripts/UI/HealthBar.cs
@@ -9,8 +9,8 @@ public class HealthBar : MonoBehaviour
 
     public void UpdateHealthBar(Character character)
     {
-        float healthPercentage = character.Health / character.characterData.maxHealth;
-        healthSlider.value = Mathf.Clamp01(healthPercentage);
+        double healthPercentage = character.Health / character.characterData.maxHealth;
+        healthSlider.value = Mathf.Clamp01((float)healthPercentage);
         healthText.text = (int)character.Health + " / " + (int)character.characterData.maxHealth;
     }
 }
